Fix navigation bug from splash screen to home screen
This bug occurred because the navigate to home screen was called several
times. Missed an else statement. Also, the Splash screen would rebuild
in the background when opening a text field. I changed this to not
happen by unsetting the splash screen from the route when we're done
with it.

Additional GUI changes.